Project/requirements.txt
Django
django-picklefield
djangorestframework
numpy
jsonfield
pandas